Implement P1 restriction as dual (transpose) of prolongation
Rather than L2-projecting and then using the Riesz map to get the
restricted residual back into the test space, we consider the coarse
basis functions to be linear combinations of the fine space basis
functions.  This gives us slightly better convergence, and as a bonus,
preserves the symmetry of preconditioned operators.